Refactor: File Convention 수정 (.js -> .jsx)
