+++ src/components/CartTrashImage.jsx
